writing size of types

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long long int i3;
 char c;
 float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(i));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(i2));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(i3));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %lu byte(s)\n", sizeof(c));
+printf("Size of a int: %lu byte(s)\n", sizeof(i));
+printf("Size of a long int: %lu byte(s)\n", sizeof(i2));
+printf("Size of a long long int: %lu byte(s)\n", sizeof(i3));
+printf("Size of a float: %lu byte(s)\n", sizeof(f));
 return (0);
 }
